# Improve MeshletMesh::from_mesh performance