@zeux 2024-06-18
Just to wrap up the METIS comments, I looked into this a little... there's indeed an RNG. The good news is that the partitioning request has an optional seed; it defaults to 4321, and can be overridden using the `set_option` API. When the partitioning begins, the seed is set from the options - so repeat invocations of the same partition API should produce the same result unless there are other sources of non-determinism.

The bad news is that the seed is used to initialize the CRT RNG. On Windows, its state is stored in a thread-local variable so everything should be fine. On Linux, its state is stored in a global that is protected with a mutex; as such, running two METIS operations concurrently, or running METIS code concurrently with anything that uses CRT srand/rand APIs, will change the results. On macOS, the state is stored in an unsynchronized global so running two operations concurrently runs into a race condition which probably reduces the quality of the random generation somewhat.

... so this explains why I saw deterministic results in practice once hash issues were resolved, but it does mean the behavior from parallel execution will likely be unusual. I don't know if METIS has other global state besides CRT RNG to worry about, let's hope it does not.