@superdump 2024-06-18
Welcome Arseny, thanks for your contribution, for meshoptimizer, and for your Niagara renderer video series. :)

We can see how things perform with the higher-level parallelisation across processing of many assets first. I do like the optimise-single-threaded-solutions-first approach. :)